+++ src/components/form/FormDocumentUploads.vue
@@ -193,6 +193,7 @@ async function onGetFile(uploadedDocument: UploadedProcessDocument) {
                 :accept="document.mimeTypes?.join(', ')"
                 :aria-required="document.isRequiredInProcess && !documentUploads.filter(obj => obj.document === document.document).length"
                 :color="fieldColor"
+                :data-test="document.acronym?.toLowerCase().replace(/_/g, '-') + '-file'"
                 :label="(document.description + (document.isRequiredInProcess ? ' *' : ''))"
                 :max-file-size="MAX_FILE_SIZE"
                 :max-files="document.isMultiple ? (MAX_FILES - documentUploads.filter(obj => obj.document === document.document).length) :
@@ -227,7 +228,7 @@ async function onGetFile(uploadedDocument: UploadedProcessDocument) {
                     </p>
                 </template>
                 <template v-slot:prepend>
-                    <QIcon name="bi-paperclip"/>
+                    <QIcon name="bi-paperclip" />
                 </template>
             </QFile>