+++ src/components/project/TableManageProjectsBtn.vue
@@ -108,10 +108,10 @@ const initOptions = () => {
 
     // Update project dates
     if ((props.projectStatus === 'PROJECT_DRAFT_PROCESSED' ||
-            props.projectStatus === 'PROJECT_PROCESSING' ||
-            props.projectStatus === 'PROJECT_VALIDATED' ||
-            props.projectStatus === 'PROJECT_REVIEW_DRAFT' ||
-            props.projectStatus === 'PROJECT_REVIEW_PROCESSING') &&
+        props.projectStatus === 'PROJECT_PROCESSING' ||
+        props.projectStatus === 'PROJECT_VALIDATED' ||
+        props.projectStatus === 'PROJECT_REVIEW_DRAFT' ||
+        props.projectStatus === 'PROJECT_REVIEW_PROCESSING') &&
         hasPerm('change_project_as_validator')) {
         options.value.push({
             icon: 'bi-calendar',
@@ -122,7 +122,7 @@ const initOptions = () => {
 
     // Change commission
     if ((props.projectStatus === 'PROJECT_PROCESSING' ||
-            props.projectStatus === 'PROJECT_VALIDATED') &&
+        props.projectStatus === 'PROJECT_VALIDATED') &&
         hasPerm('change_projectcommissionfund_as_validator')) {
         options.value.push({
             icon: 'bi-signpost',
@@ -143,7 +143,7 @@ const initOptions = () => {
 
 
     if (props.projectStatus === 'PROJECT_REVIEW_PROCESSING' || props.projectStatus === 'PROJECT_REVIEW_VALIDATED'
-        || props.projectStatus === 'PROJECT_CANCELLED') {
+        || props.projectStatus === 'PROJECT_CANCELED') {
         // View review
         options.value.push({
             icon: 'bi-eye',
@@ -243,7 +243,7 @@ async function onGetProjectFiles(projectId: number, projectName: string) {
                     @click="onOptionClick(option)"
                 >
                     <QItemSection avatar>
-                        <QAvatar :icon="option.icon"/>
+                        <QAvatar :icon="option.icon" />
                     </QItemSection>
                     <QItemSection>
                         <QItemLabel>{{ option.label }}</QItemLabel>