+++ src/components/project/TableUserProjectsBtn.vue
@@ -101,7 +101,7 @@ const initOptions = () => {
         })
     }
     if (props.projectStatus === 'PROJECT_REVIEW_PROCESSING' || props.projectStatus === 'PROJECT_REVIEW_VALIDATED'
-        || props.projectStatus === 'PROJECT_CANCELLED') {
+        || props.projectStatus === 'PROJECT_CANCELED') {
         options.value.push({
             icon: 'bi-eye',
             label: t('project.view-review'),
@@ -199,7 +199,7 @@ async function onGetProjectFiles(projectId: number, projectName: string) {
                     @click="onOptionClick(option)"
                 >
                     <QItemSection avatar>
-                        <QAvatar :icon="option.icon"/>
+                        <QAvatar :icon="option.icon" />
                     </QItemSection>
                     <QItemSection>
                         <QItemLabel>{{ option.label }}</QItemLabel>